+++ klassez/__init__.py
@@ -10,8 +10,10 @@ from pprint import pprint
 
 from . import fit, misc, sim, figures, processing, anal
 
-from .Spectra import Spectrum_1D, pSpectrum_1D, Spectrum_2D, pSpectrum_2D, Pseudo_2D, DOSY
+from .Spectra import *
 
-from .config import CM, CM_2D, COLORS, cron, safe_kws, textcolor
+from .config import CM, CM_2D, COLORS, cron, safe_kws, textcolor, _print, cprint
+
+print = cprint
 
 __version__ = "0.5a.5"